[IMP] Moved a part of permissions checks down into the business layer for /manage/cases/delete
